+++ crates/miden-lib/src/transaction/mod.rs
@@ -125,7 +125,7 @@ impl TransactionKernel {
             tx_inputs.block_header().block_num(),
         );
 
-        let mut tx_advice_inputs = TransactionAdviceInputs::new(tx_inputs, tx_args)?;
+        let mut tx_advice_inputs = TransactionAdviceInputs::new(tx_inputs, tx_args);
         if let Some(init_advice_inputs) = init_advice_inputs {
             tx_advice_inputs.extend(init_advice_inputs);
         }