+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 - Added `Note::is_network_note()` accessor (#1485).
 - [BREAKING] Update handling of the shared modules (#1490).
 - Added a new constructor for `TransactionExecutor` that accepts `ExecutionOptions` (#1502).
+- [BREAKING] `TransactionAdviceInputs` cannot return `Err` anymore (#1517).
 
 ## 0.9.5 (2025-06-20) - `miden-lib` crate only